renamed dataFilters package name to filters
